backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
drf-spectacular==0.27.1
weasyprint==60.2
Pillow==10.2.0
django-extensions==3.2.3
django-redis==5.4.0
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
requests==2.31.0

